refactor: centralize auth middleware
