+++ routes/account.js
@@ -1,12 +1,6 @@
 const express = require('express');
 const router = express.Router();
-
-function requireAuth(req, res, next) {
-  if (!req.user) {
-    return res.redirect('/login');
-  }
-  next();
-}
+const { requireAuth } = require('../middleware/auth');
 
 router.get('/', requireAuth, (req, res) => {
   res.render('account', { user: req.session.user });
